Develop

See merge request pharmadi/pharmadi-be!252

+++ src/Jhipster.Module/OrderingManagement/Module.Ordering.Infrastructure/Persistences/Repositories/CartRepository.cs
@@ -27,22 +27,28 @@ namespace Module.Factor.Infrastructure.Persistence.Repositories
         {
             try
             {
-                var obj = await _context.Carts.FirstOrDefaultAsync(i => i.UserId == request.UserId && i.ProductId == request.ProductId);
-                if (obj != null)
+                var checkPrduct = await _context.Products.FirstOrDefaultAsync(i => i.Status == 2 && i.Archived == false && i.CanOrder == true && i.Id == request.ProductId);
+                if (checkPrduct == null) throw new Exception("Sản phẩm đang đợi kiểm duyệt, không cho phép đặt hàng ");
+                else
                 {
-                    request.Id = obj.Id;
-                    request.Quantity = obj.Quantity + request.Quantity;
-                    if (request.Quantity == 0)
+                    var obj = await _context.Carts.FirstOrDefaultAsync(i => i.UserId == request.UserId && i.ProductId == request.ProductId);
+                    if (obj != null)
                     {
-                        _context.Carts.Remove(obj);
-                        return await _context.SaveChangesAsync();
-                    }
-                    obj = _mapper.Map<Cart, Cart>(request, obj);
+                        request.Id = obj.Id;
+                        request.Quantity = obj.Quantity + request.Quantity;
+                        if (request.Quantity == 0)
+                        {
+                            _context.Carts.Remove(obj);
+                            return await _context.SaveChangesAsync();
+                        }
+                        obj = _mapper.Map<Cart, Cart>(request, obj);
 
-                    return await _context.SaveChangesAsync(default);
+                        return await _context.SaveChangesAsync(default);
+                    }
+                    await _context.Carts.AddAsync(request);
+                    return await _context.SaveChangesAsync();
                 }
-                await _context.Carts.AddAsync(request);
-                return await _context.SaveChangesAsync();
+                
             }
             catch (Exception e)
             {
@@ -67,11 +73,11 @@ namespace Module.Factor.Infrastructure.Persistence.Repositories
         }
 
 
-        public async Task<ViewCartDTO> GetAllByUser(int page, int pageSize, Guid userId , int ? check)
+        public async Task<ViewCartDTO> GetAllByUser(int page, int pageSize, Guid userId, int? check)
         {
-             
+
             var view = new ViewCartDTO();
-          
+
             var res = new List<ViewCartByBrandDTO>();
             var q1 = new List<Guid>();
             if (check == null || check == 0)
@@ -99,10 +105,10 @@ namespace Module.Factor.Infrastructure.Persistence.Repositories
                 }
                 foreach (var item in query)
                 {
-                   if( item.Product.ShortName == null)
+                    if (item.Product.ShortName == null)
                     {
                         item.Product.ShortName = item.Product.ProductName.Substring(0, 25);
-                    }    
+                    }
                 }
 
                 foreach (var item in q1)
@@ -280,23 +286,83 @@ namespace Module.Factor.Infrastructure.Persistence.Repositories
         public async Task<CartResultDTO> CartResultSum(Guid userId)
         {
             var res = new CartResultDTO();
+            res.TotalPrice = 0;
+            res.TotalPayment = 0;
             var data = await _context.Carts.Include(i => i.Product).Where(i => i.UserId == userId && i.IsChoice == true).ToListAsync();
             res.Quantity = (int)data.Sum(i => i.Quantity);
 
             foreach (var item in data)
             {
-                var checkPrice = item.Product.SuggestPrice != null ? item.Product. SuggestPrice : item.Product.SalePrice;
-                var summ = (int)(item.Quantity * checkPrice);
-                res.TotalPrice += summ;
-
-                var summ1 = (int)(item.Quantity * item.Product.SalePrice);
-                res.TotalPayment += summ1;
-
-
+                var checkPrice = item.Product.SuggestPrice != null ? item.Product.SuggestPrice : item.Product.SalePrice;
+                // tính Giá theo từng sản phẩm 
+                var dis = ProductDiscount(item.UserId, item.ProductId);
+                if (dis.unit == "%")
+                {
+                    var summ = item.Quantity * item.Product.SalePrice;
+                    res.TotalPrice += summ;
+                    var summ1 = item.Quantity * item.Product.SalePrice * (decimal)(100 - dis.Discount) / 100;
+                    res.TotalPayment += summ1;
+                }
+                if (dis.unit.ToLower() == "VNĐ".ToLower())
+                {
+                    var summ = item.Quantity * item.Product.SalePrice;
+                    res.TotalPrice += summ;
+                    var summ1 = item.Quantity * item.Product.SalePrice - (decimal)(dis.Discount);
+                    res.TotalPayment += summ1;
+                }
+                if (dis.unit == "0")
+                {
+                    var summ = item.Quantity * item.Product.SalePrice;
+                    res.TotalPrice += summ;
+                    var summ1 = item.Quantity * item.Product.SalePrice;
+                    res.TotalPayment += summ1;
+                }
             }
+            res.economicalPrice = res.TotalPrice - res.TotalPayment;
             return res;
 
         }
+        private DiscountDTO ProductDiscount(Guid? userId, Guid? productId)
+        {
+            var s = new DiscountDTO();
+            var checkCart = _context.Carts.Where(i => i.UserId == userId && i.ProductId == productId).FirstOrDefault().Quantity;
+            var checkProduct = _context.productDiscounts.Where(i => i.ProductId == productId).OrderBy(i => i.Max);
+            if (checkProduct.Any())
+            {
+                foreach (var item in checkProduct)
+                {
+                    if (item.Min > checkCart)
+                    {
+                        s.Discount = 0;
+                        s.unit = "0";
+                        break;
+                    }
+                    if (item.Min <= checkCart && checkCart <= item.Max)
+                    {
+                        s.Discount = item.Discount;
+                        s.unit = item.Unit;
+                        break;
+                    }
 
+                }
+                if (s.unit == null)
+                {
+                    s.Discount = checkProduct.OrderByDescending(i => i.Max).FirstOrDefault().Discount;
+                    s.unit = checkProduct.FirstOrDefault().Unit;
+                }
+            }
+            else
+            {
+                s.Discount = 0;
+                s.unit = "0";
+            }
+            return s;
+        }
+        private class DiscountDTO
+        {
+            public float Discount { get; set; }
+
+            public string unit { get; set; }
+        }
     }
 }
